Replace getopt with optparse; tidying up elsewhere.
Added whitespace around operators.
Added a comment about the second Image being created. Only deal with
the second image if the dimensions of the resized image do not match
the desktop size.

There is currently an error when running this with the data gathered
from requests.get(). It is working properly when run against a local
file.