Ensure the image description is a single line.
While testing with a local snippet of the NASA RSS feed I found that I
had inadvertently added extra newlines and whitespace. This caused the
text output in the images to be unappealing. I'm not sure there will
ever be newlines or spaces in the actual RSS online but I may as well
clean it up just in case.